Fixed aspect ratio calculating when aspect is 0.
Fixes 2048 and others.

+++ src/main/out_video.cpp
@@ -185,8 +185,8 @@ bool video_init(const struct retro_game_geometry *geom, SDL_Window *context)
 	g_video.tex_h = geom->max_height;
 	g_video.base_w = geom->base_width;
 	g_video.base_h = geom->base_height;
-	if (geom->aspect_ratio <= 0)
-		g_video.aspect = geom->base_width / geom->base_height;
+	if (geom->aspect_ratio <= 0.0)
+		g_video.aspect = (float)geom->base_width / (float)geom->base_height;
 	else
 		g_video.aspect = geom->aspect_ratio;
